requirements.txt
websockets==16.0
uvicorn==0.40.0
python-dotenv==1.2.1
redis==7.1.0
fastapi==0.128.0
